server refactor partial
