+++ client/src/features/gatekeeper/gatekeeper-rules-reducer.js
@@ -1,5 +1,5 @@
 import { createSlice } from '@reduxjs/toolkit';
-import {queryGatekeeper} from './gatekeeper.js'
+import { queryGatekeeper } from './gatekeeper.js'
 
 export const gatekeeperRules = createSlice({
   name: 'gatekeeperRules',
@@ -14,10 +14,10 @@ export const gatekeeperRules = createSlice({
     },
 
     setDashboardResults: (state, data) => {
-        state.ruleResults = data.payload;
-      },
+      state.ruleResults = data.payload;
+    },
 
-    
+
   },
 });
 
@@ -31,30 +31,30 @@ export const selectDashboardRuleResults = state => state.gatekeeperRules.ruleRes
 
 export const populateDashboardRules = (ens) => async (dispatch, getState, axios) => {
 
-  const res = await axios.get('/dashboardRules/' + ens);
-  
-    dispatch(setDashboardRules(res.data))
+  const res = await axios.get('/dashboard/dashboardRules/' + ens);
+
+  dispatch(setDashboardRules(res.data))
 }
 
 export const applyDashboardRules = (walletAddress) => async (dispatch, getState, axios) => {
 
-    let { gatekeeperRules, connectivity, user } = getState();
+  let { gatekeeperRules, connectivity, user } = getState();
 
-    if(!connectivity.connected){
-      // wallet not connected, clear rule results
-      dispatch(setDashboardResults({}))
-    }
-    
-    else{
+  if (!connectivity.connected) {
+    // wallet not connected, clear rule results
+    dispatch(setDashboardResults({}))
+  }
+
+  else {
     let ruleTestResults = {}
     Object.entries(gatekeeperRules.dashboardRules).map(([key, value]) => {
-            ruleTestResults[key] = "";
-        
-    }) 
+      ruleTestResults[key] = "";
+
+    })
     const result = await queryGatekeeper(walletAddress, gatekeeperRules.dashboardRules, ruleTestResults, user.discord_id)
     dispatch(setDashboardResults(result))
-   }
   }
+}
